// apps/wop/app/core/layouts/types.ts
import { ReactNode } from "react"
import { Settings } from "shared-ui"

export type ContentWidth = "full" | "boxed"

export type ThemeColor = "primary" | "secondary" | "error" | "warning" | "info" | "success"

export type NavLink = {
  path?: string
  title: string
  action?: string
  subject?: string
  disabled?: boolean
  badgeContent?: string
  externalLink?: boolean
  openInNewTab?: boolean
  icon?: string | string[] | ReactNode
  badgeColor?: "default" | "primary" | "secondary" | "success" | "error" | "warning" | "info"
}

export type NavSectionTitle = {
  sectionTitle: string
  action?: string
  subject?: string
}

export type VerticalNavItemsType = (NavLink | NavSectionTitle)[]

export type LayoutProps = {
  hidden: boolean
  settings: Settings
  children: ReactNode
  verticalNavItems?: VerticalNavItemsType
  scrollToTop?: (props?: any) => ReactNode
  saveSettings: (values: Settings) => void
  footerContent?: (props?: any) => ReactNode
  verticalAppBarContent?: (props?: any) => ReactNode
}

export type BlankLayoutProps = {
  children: ReactNode
  title: String
}